requirements.txt
# Datas
Click==7.0
geopy==1.20.0
pandas==0.25.2
tabulate==0.8.3

# For simila image detection
dhash==1.3
Wand==0.5.7
Pillow==6.0.0
imutils==0.5.3
opencv-python==4.1.1.26
opencv-contrib-python==4.1.1.26
